stress-urandom: make ret argument in check_eperm call ssize_t
Don't use int, use the correct type as being passed to the call

+++ stress-urandom.c
@@ -35,7 +35,7 @@ static const stress_help_t help[] = {
     defined(RNDADDTOENTCNT) ||	\
     defined(RNDRESEEDCRNG) || 	\
     defined(__linux__)
-static void check_eperm(const stress_args_t *args, const int ret, const int err)
+static void check_eperm(const stress_args_t *args, const ssize_t ret, const int err)
 {
 	if ((g_opt_flags & OPT_FLAGS_VERIFY) &&
 	    ((ret == 0) || ((err != EPERM) && (err != EINVAL) && (err != ENOTTY)))) {
